Provide secondary contructor for SafeToken, implements #137
Changing the argument of the primary ctor from String to CharSequence may break some libraries which still depend on the older version.
This commit aims to fix this by providing the old constructor as a seconday constructor.

In addition this commit updates mockito and removes jmockit for some issues caused by them on CI with Java 11.